day3 part2

+++ aoc_2018/src/day1.rs
@@ -2,17 +2,17 @@
 use std::collections::HashSet;
 
 #[aoc_generator(day1)]
-pub fn input_frequencies(input: &str) -> Vec<i32> {
+pub fn input_frequencies(input: &str) -> Vec<isize> {
     input.lines().map(|l| l.parse().unwrap()).collect()
 }
 
 #[aoc(day1, part1)]
-pub fn part1(input: &[i32]) -> i32 {
+pub fn part1(input: &[isize]) -> isize {
     input.iter().sum()
 }
 
 #[aoc(day1, part2)]
-pub fn part2(input: &[i32]) -> i32 {
+pub fn part2(input: &[isize]) -> isize {
     let mut freq = 0;
     let mut frequencies = HashSet::new();
     frequencies.insert(freq);
